refactor(credit): tidy up engine building
